+++ compiler/src/config.rs
@@ -2,6 +2,7 @@ use derive_more::{Deref, derive::From};
 use serde::Deserialize;
 
 use crate::CONFIG_ENV_PREFIX;
+use crate::passes::{InstrumentationRules, InternalizationRules};
 use common::log_info;
 
 #[derive(Debug, Default, Clone, Deserialize)]
@@ -98,234 +99,6 @@ impl Default for RuntimeShimLocation {
 fn default_runtime_shim_crate_name() -> String {
     "leaf".to_string()
 }
-
-pub(crate) type InstrumentationRules = InclusionRules<EntityFilter>;
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(tag = "entity")]
-#[serde(rename_all = "snake_case")]
-pub(crate) enum EntityFilter {
-    #[serde(alias = "body")]
-    WholeBody(WholeBodyFilter),
-    #[serde(alias = "dyn_def")]
-    MethodDynDefinition(MethodDynDefinitionFilter),
-    PlaceInfo(PlaceInfoFilter),
-    OperandInfo(OperandInfoFilter),
-    #[serde(alias = "const_type")]
-    ConstantType(ConstantTypeFilter),
-    Assignment(AssignmentFilter),
-    AssignmentInfo(AssignmentFilter),
-    StorageLifetime(StorageLifetimeMarkerFilter),
-    CallFlow(CallFlowFilter),
-    Drop(DropFilter),
-    Switch(SwitchFilter),
-}
-
-#[derive(Debug, Clone, Deserialize, From)]
-pub(crate) struct WholeBodyFilter(pub(crate) LogicFormula<EntityLocationFilter>);
-
-#[derive(Debug, Clone, Deserialize)]
-pub(crate) struct MethodDynDefinitionFilter(pub(crate) LogicFormula<EntityLocationFilter>);
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(tag = "kind")]
-#[serde(rename_all = "snake_case")]
-pub(crate) enum PlaceInfoFilter {
-    #[serde(alias = "struct")]
-    Structure(PlaceInfoStructureFilter),
-    #[serde(alias = "addr")]
-    Address(PlaceAddressFilter),
-    #[serde(alias = "ty")]
-    Type(PlaceTypeFilter),
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(tag = "kind")]
-#[serde(rename_all = "snake_case")]
-pub(crate) struct StorageLifetimeMarkerFilter {
-    pub(crate) kind: LogicFormula<StorageLifetimeMarker>,
-    #[serde(flatten)]
-    pub(crate) loc: LogicFormula<StorageLifetimeLocationFilter>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub(crate) enum EntityLocationFilter {
-    #[serde(alias = "def_path")]
-    DefPathMatch(PatternMatch),
-    Crate(CrateFilter),
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub(crate) enum CrateFilter {
-    #[serde(alias = "is_external")]
-    Externality(bool),
-    Name(PatternMatch),
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(tag = "kind")]
-#[serde(rename_all = "snake_case")]
-pub(crate) struct PlaceInfoStructureFilter {
-    pub(crate) piece: LogicFormula<PlaceStructurePiece>,
-    #[serde(flatten)]
-    pub(crate) loc: LogicFormula<EntityLocationFilter>,
-}
-
-#[derive(Debug, Clone, Copy, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub(crate) enum PlaceStructurePiece {
-    Local,
-    Deref,
-    Field,
-    Index,
-    ConstantIndex,
-    Subslice,
-    Downcast,
-    OpaqueCast,
-    UnwrapUnsafeBinder,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-pub(crate) struct PlaceAddressFilter(pub(crate) LogicFormula<EntityLocationFilter>);
-
-#[derive(Debug, Clone, Deserialize)]
-pub(crate) struct PlaceTypeFilter(pub(crate) LogicFormula<EntityLocationFilter>);
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(tag = "kind")]
-#[serde(rename_all = "snake_case")]
-pub(crate) struct OperandInfoFilter {
-    pub(crate) kind: LogicFormula<OperandKind>,
-    #[serde(flatten)]
-    pub(crate) loc: OperandInfoLocationFilter,
-}
-
-#[derive(Debug, Clone, Copy, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub(crate) enum OperandKind {
-    Copy,
-    Move,
-    #[serde(alias = "const")]
-    Constant,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-pub(crate) struct ConstantTypeFilter {
-    #[serde(alias = "type")]
-    pub(crate) ty: LogicFormula<ConstantType>,
-    #[serde(flatten)]
-    pub(crate) loc: OperandInfoLocationFilter,
-}
-
-#[derive(Debug, Clone, Copy, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub(crate) enum ConstantType {
-    Bool,
-    Char,
-    Int,
-    Float,
-    Str,
-    ByteStr,
-    Ptr,
-    Zst,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-pub(crate) struct OperandInfoLocationFilter(pub(crate) LogicFormula<EntityLocationFilter>);
-
-#[derive(Debug, Clone, Deserialize)]
-pub(crate) struct AssignmentFilter {
-    pub(crate) kind: LogicFormula<AssignmentKind>,
-    #[serde(flatten)]
-    pub(crate) loc: AssignmentLocationFilter,
-}
-
-#[derive(Debug, Clone, Copy, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub(crate) enum AssignmentKind {
-    Use,
-    Repeat,
-    Ref,
-    ThreadLocalRef,
-    RawPtr,
-    Cast,
-    BinaryOp,
-    UnaryOp,
-    Discriminant,
-    Aggregate,
-    ShallowInitBox,
-    WrapUnsafeBinder,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-pub(crate) struct AssignmentLocationFilter(pub(crate) LogicFormula<EntityLocationFilter>);
-
-#[derive(Debug, Clone, Copy, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub(crate) enum StorageLifetimeMarker {
-    Live,
-    Dead,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-pub(crate) struct StorageLifetimeLocationFilter(pub(crate) LogicFormula<EntityLocationFilter>);
-
-#[derive(Debug, Clone, Deserialize)]
-pub(crate) struct CallFlowFilter {
-    pub(crate) kind: LogicFormula<CallFlowPartKind>,
-    #[serde(flatten)]
-    pub(crate) loc: CallFlowLocationFilter,
-}
-
-#[derive(Debug, Clone, Copy, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub(crate) enum CallFlowPartKind {
-    CallControl,
-    CallInput,
-    #[serde(alias = "func_data")]
-    FunctionData,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-pub(crate) struct CallFlowLocationFilter(pub(crate) LogicFormula<EntityLocationFilter>);
-
-#[derive(Debug, Clone, Deserialize)]
-pub(crate) struct DropFilter {
-    pub(crate) kind: LogicFormula<DropPartKind>,
-    #[serde(flatten)]
-    pub(crate) loc: DropLocationFilter,
-}
-
-#[derive(Debug, Clone, Copy, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub(crate) enum DropPartKind {
-    CallControl,
-    CallInput,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-pub(crate) struct DropLocationFilter(pub(crate) LogicFormula<EntityLocationFilter>);
-
-#[derive(Debug, Clone, Deserialize)]
-pub(crate) struct SwitchFilter {
-    pub(crate) kind: LogicFormula<SwitchPartKind>,
-    #[serde(flatten)]
-    pub(crate) loc: SwitchLocationFilter,
-}
-
-#[derive(Debug, Clone, Copy, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub(crate) enum SwitchPartKind {
-    Control,
-    Data,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-pub(crate) struct SwitchLocationFilter(pub(crate) LogicFormula<EntityLocationFilter>);
-
-pub(crate) type InternalizationRules = InclusionRules<LogicFormula<PatternMatch>>;
 
 const CONFIG_FILENAME: &str = "leafc_config";
